fix product errors and create handler exception for Feign Client

+++ order-service/src/main/java/com/example/order_service/dto/request/OrderCreationRequest.java
@@ -12,6 +12,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -31,8 +32,7 @@ public class OrderCreationRequest implements Serializable {
     @JsonFormat(pattern = AppConstant.LOCAL_DATE_TIME_FORMAT, shape = JsonFormat.Shape.STRING)
     @DateTimeFormat(pattern = AppConstant.LOCAL_DATE_TIME_FORMAT)
     LocalDateTime orderDate;
-
     String orderDesc;
-    Double orderFee;
+    BigDecimal orderFee;
     List<OrderItemRequest> items;
 }
